⚒ BREAKING CHANGES! Refactored pagination and dropdown components
