+++ app/components/characters/FilterDropdown/styles/StyledCheckbox.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const StyledCheckbox = styled.input.attrs({ type: 'checkbox' })`
+	filter: hue-rotate(135deg) brightness(0.8);
+`
