+++ app/components/characters/FilterDropdown/styles/index.tsx
@@ -1,0 +1,2 @@
+export * from './StyledCheckbox'
+export * from './StyledDropdown'
