+++ app/components/characters/Pagination/styles/StyledPaginationItem.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const StyledPaginationItem = styled.button`
+	background: linear-gradient(112.71deg, #ab1818 17.21%, #520000 91.53%);
+	box-shadow: 0px 20px 40px rgba(0, 0, 0, 0.2);
+	color: white;
+`
