+++ app/components/characters/Pagination/styles/index.tsx
@@ -1,0 +1,1 @@
+export * from './StyledPaginationItem'
