+++ app/components/characters/Pagination/utils/pageNumbers.tsx
@@ -1,0 +1,1 @@
+export const pages = [1, 2, 3, 4, 5, 6, 7, 8, 9]
