+++ app/components/characters/SortDropdown/styles/StyledRadio.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const StyledRadio = styled.input.attrs({ type: 'radio' })`
+	filter: hue-rotate(140deg) brightness(0.8);
+`
