+++ app/components/characters/SortDropdown/styles/index.tsx
@@ -1,0 +1,1 @@
+export * from './StyledRadio'
